pyproject.toml
[tool.poetry]
name = "sister_bot"
version = "1.0.0"
description = "基于 Nonebot2 和 go-cqhttp 开发的QQ机器人，不正经的妹妹"
authors = ["ITSevin <itsevin@qq.com>"]
license = "AGPL-3.0"
readme = "README.md"
homepage = "https://github.com/itsevin/sister_bot"
repository = "https://github.com/itsevin/sister_bot"
documentation = "https://github.com/itsevin/sister_bot#readme"
keywords = ["nonebot", "nonebot2", "qqbot", "gocqhttp", "bot"]

[[tool.poetry.source]]
name = "ali"
default = true
url = "https://mirrors.aliyun.com/pypi/simple/"

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = "^2.0.0rc4"
httpx = "^0.23.0"
nonebot-plugin-apscheduler = "^0.2.0"
nonebot-plugin-htmlrender = "^0.2.0"
nonebot-plugin-imageutils = "^0.1.13.4"
nonebot-adapter-onebot = "^2.2.0"
nonebot-plugin-handle = "^0.1.6"
nonebot-plugin-memes = "^0.3.9"
nonebot-plugin-minesweeper = "^0.1.8"
nonebot-plugin-remake = "^0.2.7"
nonebot-plugin-savor = "^0.2.1"
nonebot-plugin-simplemusic = "^0.1.5"
nonebot-plugin-tarot = "0.4.0.post4"
nonebot-plugin-wordle = "^0.1.11"
nonebot-plugin-record = "^1.0.2"
nonebot-plugin-today-in-history = "0.1.1"
nonebot-plugin-read-60s = "^0.2.4"
nonebot-plugin-boardgame = "^0.1.8"
nonebot-plugin-maze = "^0.2.4"
nonebot-plugin-withdraw = "^0.2.3"
nonebot-plugin-code = "^0.0.4"
nonebot-plugin-heweather = "^0.6.0"
mcstatus = "^10.0.1"
yinglish = "^1.0.1"
config = "^0.5.1"
aiohttp = "^3.8.3"
nonebot-plugin-manager = "^0.5.14"
nonebot-plugin-bing-chat = "^0.7.2"
nonebot-plugin-naturel-gpt = "^2.0.4"
ujson = "^5.8.0"
ruamel-yaml = "^0.17.32"
poetry = "^1.5.1"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_apscheduler", "nonebot_plugin_htmlrender", "nonebot_plugin_imageutils", "nonebot_plugin_handle", "nonebot_plugin_memes", "nonebot_plugin_minesweeper", "nonebot_plugin_read_60s", "nonebot_plugin_remake", "nonebot_plugin_savor", "nonebot_plugin_simplemusic", "nonebot_plugin_tarot", "nonebot_plugin_today_in_history", "nonebot_plugin_wordle", "nonebot_plugin_record", "nonebot_plugin_boardgame",  "nonebot_plugin_maze", "nonebot_plugin_withdraw", "nonebot_plugin_code", "nonebot_plugin_heweather", "nonebot_plugin_naturel_gpt", "nonebot_plugin_bing_chat"]
plugin_dirs = ["plugins/basic_plugins", "plugins/plugins", "plugins/extra_plugins"]
builtin_plugins = []

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"